+++ CHANGELOG.md
@@ -19,7 +19,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 
-## [2.9.0] - 2019-08-07
+## [2.12.0] - 2019-09-26
+
+### Security
+- dependency updates
+
+### Added
+- Smart.Transactions: Offline functionality (cache & offline receipt)
+- Smart.Ident: display placeholder page
+
+
+## [2.11.0] - 2019-08-07
 
 ### Security
 - dependency updates
@@ -413,3 +423,4 @@ Internal developer release
 [2.9.0]:https://github.com/secucard/secucard-connect-java-sdk/compare/v2.8.0...v2.9.0
 [2.10.0]:https://github.com/secucard/secucard-connect-java-sdk/compare/v2.9.0...v2.10.0
 [2.11.0]:https://github.com/secucard/secucard-connect-java-sdk/compare/v2.10.0...v2.11.0
+[2.12.0]:https://github.com/secucard/secucard-connect-java-sdk/compare/v2.11.0...v2.12.0